modification de la bdd + fix bug photo de profil

+++ backend/controllers/User.js
@@ -5,16 +5,19 @@ const jwt = require("jsonwebtoken");
 const fs = require("fs");
 
 exports.signup = (req, res) => {
-
   const password = req.body.password;
-  const regexPassword = new RegExp(/^(?=.*[a-z])(?=.*[A-Z])(?=.*[0-9])(?=.*[!@#\$%\^&\*])(?=.{8,})/);
+  const regexPassword = new RegExp(
+    /^(?=.*[a-z])(?=.*[A-Z])(?=.*[0-9])(?=.*[!@#\$%\^&\*])(?=.{8,})/
+  );
   const testPassword = regexPassword.test(password);
 
-  const email = req.body.email; 
-  const regexEmail = new RegExp(/[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*@(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?/);
+  const email = req.body.email;
+  const regexEmail = new RegExp(
+    /[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*@(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?/
+  );
   const testEmail = regexEmail.test(email);
 
-  if(testEmail && testPassword) {
+  if (testEmail && testPassword) {
     User.findOne({ where: { email: req.body.email } }).then((user) => {
       if (!user) {
         bcrypt
@@ -25,7 +28,7 @@ exports.signup = (req, res) => {
               firstName: req.body.firstName,
               email: req.body.email,
               password: hash,
-              picture: `http://localhost:3000/images/default_profil_picture.png`,
+              picture: `http://localhost:3000/images/default_profil_picture.jpg`,
               isAdmin: false,
             });
           })
@@ -43,8 +46,8 @@ exports.signup = (req, res) => {
       }
     });
   } else {
-    const message = "Mot de passe ou email incorrect"
-    return res.status(400).json({message})
+    const message = "Mot de passe ou email incorrect";
+    return res.status(400).json({ message });
   }
 };
 
@@ -63,7 +66,7 @@ exports.login = (req, res) => {
           return res.status(401).json({ message });
         }
         const token = jwt.sign(
-          { id: user.id, isAdmin: user.isAdmin,  },
+          { id: user.id, isAdmin: user.isAdmin },
           `Mon_token_secret`,
           {
             expiresIn: "86400000",
@@ -90,7 +93,7 @@ exports.login = (req, res) => {
 
 exports.getAllUsers = (req, res) => {
   User.findAll({
-    attributes : ["id",'name',"firstName", 'description', 'picture']
+    attributes: ["id", "name", "firstName", "description", "picture"],
   })
     .then((users) => {
       const message = `la liste des utilisateurs a bien été récupérée . `;
@@ -103,7 +106,6 @@ exports.getAllUsers = (req, res) => {
 };
 
 exports.getOneUser = (req, res) => {
-  
   User.findOne({
     where: { id: req.params.id },
   }).then((user) => {
@@ -112,7 +114,7 @@ exports.getOneUser = (req, res) => {
       return res.status(400).json({ message });
     }
     const userData = {
-      id:user.id,
+      id: user.id,
       email: user.email,
       firstName: user.firstName,
       name: user.name,
@@ -144,6 +146,18 @@ exports.updateProfil = (req, res) => {
     if (!user) {
       const message = "L'utilisateur demandé n'existe pas .";
       return res.status(404).json({ message });
+    }
+    if (
+      user.picture === `http://localhost:3000/images/default_profil_picture.jpg`
+    ) {
+      user
+        .update(userObject, {
+          where: { id: userId },
+        })
+        .catch((err) => {
+          const message = `Impossible de modifier le profil, veuillez réessayer ultérieurement. `;
+          res.status(500).json({ message, err });
+        });
     } else {
       const filename = user.picture.split("/images/")[1];
       fs.unlink(`images/${filename}`, () => {
@@ -156,9 +170,9 @@ exports.updateProfil = (req, res) => {
             res.status(500).json({ message, err });
           });
       });
-      const message = `L'utilisateur ${user.name} ${user.firstName} a bien été modifié`;
-      res.json({ message, data: userObject });
     }
+    const message = `L'utilisateur ${user.name} ${user.firstName} a bien été modifié`;
+    res.json({ message, data: userObject });
   });
 };
 
@@ -174,8 +188,10 @@ exports.deleteUser = (req, res) => {
         const message = "L'utilisateur demandé n'existe pas .";
         return res.status(404).json({ message });
       }
-      const filename = user.picture.split("/images/")[1];
-      fs.unlink(`images/${filename}`, () => {
+      if (
+        user.picture ===
+        `http://localhost:3000/images/default_profil_picture.jpg`
+      ) {
         User.destroy({ where: { id: user.id } })
           .then((_) => {
             const message = `L'utilisateur ${user.name} ${user.firstName} a bien été supprimé .`;
@@ -185,7 +201,20 @@ exports.deleteUser = (req, res) => {
             const message = `Impossible de supprimer le profil, veuillez réessayer ultérieurement. `;
             res.status(500).json({ message, err });
           });
-      });
+      } else {
+        const filename = user.picture.split("/images/")[1];
+        fs.unlink(`images/${filename}`, () => {
+          User.destroy({ where: { id: user.id } })
+            .then((_) => {
+              const message = `L'utilisateur ${user.name} ${user.firstName} a bien été supprimé .`;
+              res.json({ message, data: user });
+            })
+            .catch((err) => {
+              const message = `Impossible de supprimer le profil, veuillez réessayer ultérieurement. `;
+              res.status(500).json({ message, err });
+            });
+        });
+      }
     });
   } else {
     const message = `Vous n'êtes pas autorisés à faire cette action`;
@@ -194,29 +223,26 @@ exports.deleteUser = (req, res) => {
 };
 
 exports.authentification = (req, res) => {
-    
   const token = req.cookies.jwt;
   const decodedToken = jwt.verify(token, `Mon_token_secret`);
   const userId = decodedToken.id;
   const userAdmin = decodedToken.isAdmin;
 
-  if (token && token!=null) {
-      res.status(200).json({id:userId, isAdmin:userAdmin});
+  if (token && token != null) {
+    res.status(200).json({ id: userId, isAdmin: userAdmin });
   } else {
-      res.status(400).json('Token innexistant !');
-  };
-
+    res.status(400).json("Token innexistant !");
+  }
 };
 
-exports.logout = (req,res) => {
-  
+exports.logout = (req, res) => {
   const token = req.cookies.jwt;
 
   if (token) {
-      res.cookie('jwt', '', { maxAge: 1 });
-      res.redirect('/');
+    res.cookie("jwt", "", { maxAge: 1 });
+    res.redirect("/");
   } else {
-    const message = `Vous n'êtes pas connecté`
-      return res.status(401).json({ message });
-  };
-}
+    const message = `Vous n'êtes pas connecté`;
+    return res.status(401).json({ message });
+  }
+};
